update to ACK runtime v0.13.0 (#27)
Updates to the ACK runtime v0.13.0 release. Only changes other than the
go.mod dependency change were adding a number of lines to generator.yaml
in order to instruct the latest code generator how to deal with some of
the RDS API's ReadMany operations that have a singular primary
identifier field in the Input shape.

By submitting this pull request, I confirm that my contribution is made under
the terms of the Apache 2.0 license.