Ensure Synced Condition on DB instance/cluster (#45)
Adds tests that the ACK.ResourceSynced Condition is present on resources
of kind DBCluster and DBInstance and that the status of those conditions
are either False or True depending on the expected state transition of
the DB instance or cluster.

This involved creating some helper functions in test/e2e/condition.py
that assert that a Condition of a specific type and status is present in
a resource's Status.Conditions collection.

In grep'ing through logs for the rds-controller, I noticed that the
AvailabilityZone field for the DBInstance should be late-initialized and
so added that to the generator.yaml file.

Finally, found that ACK.ResourceSynced/True was not being set on
DBCluster and DBInstance resources when the status of the instance or
cluster made its way to 'available'. I added an else block to the
generic hook code templates to ensure that ACK.ResourceSynced condition
with a status of True is present when the resource's status transitions
to 'available'.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>

By submitting this pull request, I confirm that my contribution is made under
the terms of the Apache 2.0 license.