Add check for DB cluster available before updating (#56)
Fixes: https://github.com/aws-controllers-k8s/community/issues/1107

Description of changes:
If a DB cluster is in any state other than `Available`, then an attempt to update it will return a terminal `InvalidDBClusterStateFault` error. This error will stop the cluster from reconciling any further, even though the underlying RDS DB cluster is working properly.

This PR checks that the DB cluster is in the `Available` state before running the custom update code, otherwise re-queuing it until it is.

By submitting this pull request, I confirm that my contribution is made under
the terms of the Apache 2.0 license.