Fix `DBInstance` for updating PI and add tests (#95)
Fixes https://github.com/aws-controllers-k8s/community/issues/1346

Description of changes:
Updating `EnablePerformanceInsights` caused the `DBInstance` to cycle continuously, as it tried to reconcile the server side default values and handle the asynchronous status updates. Rather than relying on a requeue error to wait for the status to update, this pull request modifies the `DBInstance` custom code to set `ResourceSynced = false` on the resource (for the runtime logic to cause the requeue instead). Furthermore, we remove the server side defaulted values from the diff if they have not been updated in the desired spec.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.