support overriding parameter values on param group
Adds code to allow a user to specify a set of parameter override values
for a DB parameter group.

The `Spec.ParameterOverrides` is a `map[string]*string` containing the
parameter names and values for parameters in a DB parameter group that
the user wishes to override from their defaults.

The `Status.ParameterOverrideStatuses` field is a slice of `Parameter`
structs that contains more "status-y" information about those overridden
parameters.

Still yet to work out is how I'm going to keep track of which parameters
for which DB engine family can have ApplyMethod=intermediate and which
ones must have ApplyMethod=pending-reboot.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>