add parameter metadata lookup to param handling
Adds a cache to the DB parameter group hooks.go file that stores
metadata about engine family parameter defaults. This metadata informs
the resource manager whether a particular parameter is modifiable and
whether changes to it can be applied immediately or require a reboot of
the DB instance.

The cache is a simple one, with no support for TTL or expiry and
because there is nowhere to "attach" the cache that has access to the
SDK API object (in order to call RDS `DescribeEngineDefaultParameters`
API), the cache's `get` method has to have a "fetcher" function pointer
passed into it. This isn't super clean but there isn't currently a way
to, say, add the cache as a member of the `AWSResourceManager` struct
since that is all code-generated and this is a specific to the DB
Parameter Group resource manager.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>