Update AWS SDK to v1.44.174
This brings in the latest RDS API changes, including the ability
to interface directly with Secrets Manager when creating a new
DBInstance or DBCluster.

There is some follow-up work to do, i.e. determining how to
integrate support for BlueGreenDeployment, but at a minimum this
makes the RDS controller current with the existing APIs.