Copy the security groups from status into spec
Because the names of the security group fields returned by the API
differ the `VPCSecurityGroupsIDs` is not read from the API response into
the _latest_. Instead the values of the _desired_ is copied. This means
that the controller determines that there is not need for an update.
Instead by copying the the VPCSecurityGroups field from the status into
the spec the controller can correctly determine that there is a
difference between the latest and desired. This operation is only
performed when a Security Group has been specified. This affects both
type __DBCluster__ and __DBInstance__.