Add asynchronous status checks for DBProxy (#137)
The DBProxy APIs put the resource into `creating` state for ~10 minutes before becoming `available`, as well as `deleting` before it is gone. This PR copies the systems we use for `DBInstance`, which has the same properties, into custom hooks for `DBProxy`. It also updates the tests for `DBProxy` to use proper fixturing and to use a autogenerated resource name suffix (so tests don't collide).