Copy the parameter groups from status into spec (#141)
Issue https://github.com/aws-controllers-k8s/community/issues/1744, if available:

**Description of changes:**
Copying the the parameter group fields from the status into the spec so that the controller can correctly determine that there is a difference between the latest and desired. This affects both type `DBCluster` and `DBInstance`.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.