prevent DB Instance desired field state flapping (#151)
Both ModifyDBInstance and DescribeDBInstances returns a DBInstance struct that contains the *previously set* values for various mutable fields. This is problematic because it causes a "flopping" behaviour when the user has modified a Spec field from value A to value B but the output shape from ModifyDBInstance for that field contains value A, the standard SetOutput Go code generated above will set the Spec field to the *old* value again. The next time the reconciler runs, it will read the latest observed resource, see a difference between the desired and the latest state (that actually does not exist because the difference is comparing the value of the fields before they were changed) and attempt to modify the field from value B to value A again, causing a flop loop.

Luckily, the Output shape's DBInstance struct contains a `PendingModifiedValues` struct which contains those field values that the user specified. So, we can use these to "reset" the Spec back to the appropriate user-specified values.

This commit does exactly this. It looks in the `PendingModifiedValues` struct for any non-nil field that matches a field in the `Spec` and if it finds a match, sets the `Spec` field to the value in `PendingModifiedValues`. This solves the flapping problem for all of the fields in `PendingModifiedValues`, including `DBInstanceClass`, `AllocatedStorage`, `MultiAZ` and `StorageType`.

Fixes Issue aws-controllers-k8s/community#1773
Fixes Issue aws-controllers-k8s/community#1716
Fixes Issue aws-controllers-k8s/community#1376

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.