Support `DBInstance` and `DBCluster` MasterUserPassword updates
This patch adds a new feature to the rds-controller allowing users to
update their `DBInstance` and `DBCluster` passwords, by modifying the
secretReference (namespace, name or key).

Prior to this patch the controller wasn't able to update `DBInstance`
and `DBCluster` passwords due to it's inability to calculate the
difference between a desired and observed states. The fact that the
API doesn't return any information about the password combined with the
controller unable to detect changes of the secret refs

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>