DBInstance/DBCluster password updates on secret ref changes (#170)
This patch adds a new feature to the rds-controller allowing users to
update their `DBInstance` and `DBCluster` passwords, by modifying the
secretReference (namespace, name or key).

Prior to this patch the controller wasn't able to update `DBInstance`
and `DBCluster` passwords due to it's inability to calculate the
difference between a desired and observed states. The fact that the
API doesn't return any information about the password combined with the
controller unable to detect changes of the secret refs


By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.