fix storageType delta (#210)
Fixes:
``` When the StorageType parameter for a DBCluster is explicitly set to its default value ("aurora"), the AWS RDS API doesn't return this field in the response. However, when set to a non-default value (like "aurora-iopt1"), the API does return the field. This inconsistency causes problems with the controller's delta comparison logic.```

Description of changes:
Modified the `delta_pre_compare` hook to default the StorageType field to "aurora" when it's nil in the latest resource state. This addresses the issue where the AWS RDS API doesn't return the StorageType field when it's set to the default value "aurora".

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.