fix: remove SubnetIDs duplicates
In sdkFind hook, the SubnetIDs returned by AWS are appended
to the latest resource Spec.SubnetIDs. This ended up being an
issue because the latest resource is a deep copy of desired.
And there were unnecessary uedate calls were being made due
to this issue.

This change ensures the only the SubnetIDs returned by the
DescribeDBSubnetGroup call are populated in the latest SubnetIDs.