fix: remove SubnetIDs duplicates (#216)
Issue [#2083](https://github.com/aws-controllers-k8s/community/issues/2083)

Description of changes:
In sdkFind hook, the SubnetIDs returned by AWS are appended 
to the latest resource Spec.SubnetIDs. This ended up being an 
issue because the latest resource is a deep copy of desired. 
And there were unnecessary update calls were being made due 
to this issue.

This change ensures the only the SubnetIDs returned by the 
DescribeDBSubnetGroup call are populated in the latest SubnetIDs.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.