fix: ignore field setting in sdkUpdate DBInstance
When we make a ModifyDBInstance call, the output returned is outdated.
Currently we've been relying on `pendingModifiedValues` to return the
desired field values, but based on the longs, pendingModifiedValues is
not a reliable source.

These changes specifically ignore setting the fields after
ModifyDBInstance is called.

A better approach can be to ignore setting anything after Update is
called. Just in case we're missing some values that we should ignore
instead!!!