fix: omit unchanged fields on update && custom delta for dbcluster fields (#262)
Issue #, if available:

Description of changes:
With these changes, we ensure we call ModifyDBInstance only with the
changed fields.
Another change added to this PR is ensuring we check for deltas on
certain fields only when DBClusterIdentifier is nil. This is because
these fields can only be modified by the DBCluster.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.